HW2/stockanalysis/requirements.txt
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blis==1.0.2
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
frozendict==2.4.6
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
murmurhash==1.0.11
numpy==2.0.2
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
panda==0.3.1
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
preshed==3.0.9
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2024.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.0
ta==0.11.0
thinc==8.3.2
threadpoolctl==3.5.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
xgboost==2.1.3
XlsxWriter==3.2.0
yfinance==0.2.51
